requirements.txt
# Telegram Bot
aiogram==3.22.0
aiohttp==3.10.10

# ML and Computer Vision
torch==2.5.1
torchvision==0.20.1
timm==1.0.11
scikit-learn==1.5.2
pandas==2.2.3
opencv-python==4.10.0.84
Pillow==11.0.0

# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Configuration and Environment
python-dotenv==1.0.1
pydantic==2.11.10
pydantic-settings==2.11

# Logging
loguru==0.7.3

# Utilities
httpx==0.27.2
tqdm==4.66.5
